refactor: exception directory move
