+++ src/main/java/nextstep/api/favorite/domain/service/SimpleFavoriteService.java
@@ -13,7 +13,7 @@ import nextstep.api.favorite.domain.model.dto.outport.FavoriteInfo;
 import nextstep.api.favorite.domain.model.entity.Favorite;
 import nextstep.api.favorite.infrastructure.FavoriteRepository;
 import nextstep.common.exception.favorite.FavoriteNotFoundException;
-import nextstep.common.exception.member.AuthenticationException;
+import nextstep.common.exception.auth.AuthenticationException;
 
 @Service
 @RequiredArgsConstructor
