+++ src/main/java/nextstep/common/exception/auth/AuthorizationException.java
@@ -1,4 +1,4 @@
-package nextstep.common.exception.member;
+package nextstep.common.exception.auth;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
